Move preview cache entry to shared module
