+++ src/omym/features/metadata/usecases/ports.py
@@ -7,10 +7,11 @@ Why: Decouple use cases from concrete DB and cache adapters for testing and swap
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from pathlib import Path
 from sqlite3 import Connection
 from typing import Protocol, runtime_checkable
+
+from omym.shared.previews import PreviewCacheEntry
 
 
 @runtime_checkable
@@ -82,17 +83,6 @@ class ArtistCachePort(Protocol):
     def clear_cache(self) -> bool:
         """Erase cached artist data."""
         ...
-
-
-@dataclass(frozen=True, slots=True)
-class PreviewCacheEntry:
-    """Value object describing a cached dry-run preview."""
-
-    file_hash: str
-    source_path: Path
-    base_path: Path
-    target_path: Path | None
-    payload: dict[str, object]
 
 
 @runtime_checkable
