+++ src/omym/shared/__init__.py
@@ -1,5 +1,6 @@
 """Shared cross-cutting utilities exposed at the package level."""
 
 from .path_components import ComponentValue
+from .previews import PreviewCacheEntry
 
-__all__ = ["ComponentValue"]
+__all__ = ["ComponentValue", "PreviewCacheEntry"]
