+++ tests/features/metadata/test_music_file_processor.py
@@ -19,7 +19,7 @@ from omym.features.metadata import (
     ProcessingEvent,
     ProcessResult,
 )
-from omym.features.metadata.usecases.ports import PreviewCacheEntry
+from omym.shared.previews import PreviewCacheEntry
 
 
 @pytest.fixture
